requirements.txt
aiohttp
BingImageCreator-fork>=0.7.6
certifi
httpx
prompt_toolkit
requests
rich
curl_cffi
setuptools
pytest